Fix import issue caused by moving components
