+++ client/src/components/App.js
@@ -11,7 +11,7 @@ import TopBar from "./layout/TopBar";
 import PizzaPlaceList from "./PizzaPlaceList.js";
 import PizzaPlaceShow from "./PizzaPlaceShow.js";
 import NewPizzaPlaceForm from "./NewPizzaPlaceForm.js";
-import ChangeUserImage from "../uploads/ChangeUserImage";
+import ChangeUserImage from "./uploads/ChangeUserImage"
 
 const App = (props) => {
   const [currentUser, setCurrentUser] = useState(undefined);
