+++ client/src/components/registration/RegistrationForm.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import FormError from "../layout/FormError";
 import config from "../../config";
-import UploadUserImage from "../../uploads/UploadUserImage";
+import UploadUserImage from "../uploads/UploadUserImage";
 
 const RegistrationForm = () => {
   const [userPayload, setUserPayload] = useState({
